+++ Waka/canoe.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "boat.h"
+
+class canoe
+{
+	boat boat_{};
+
+public:
+
+};
