+++ Waka/monohull.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "hull.h"
+
+class monohull
+{
+private:
+	hull hull_{};
+
+public:
+
+};
