+++ Waka/multihull.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "hull.h"
+
+class multihull
+{
+private:
+	hull hull_{};
+
+public:
+
+};
