+++ Waka/paddle.cpp
@@ -1,0 +1,7 @@
+#include "stdafx.h"
+#include "paddle.h"
+
+const int paddle::getSpeed()
+{
+	return 0;
+}
