+++ Waka/propulsion.cpp
@@ -1,18 +1,7 @@
 #include "stdafx.h"
 #include "propulsion.h"
-#include <cassert>
 
 const int propulsion::getSpeed()
-{
-	return 0;
-}
-
-const int paddle::getSpeed()
-{
-	return 0;
-}
-
-const int sail::getSpeed()
 {
 	return 0;
 }
