+++ Waka/raft.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "boat.h"
+
+class raft
+{
+	boat boat_{};
+
+public:
+
+};
