+++ Waka/sail.cpp
@@ -1,0 +1,7 @@
+#include "stdafx.h"
+#include "sail.h"
+
+const int sail::getSpeed()
+{
+	return 0;
+}
