+++ Waka/sailboat.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "boat.h"
+
+class sailboat
+{
+	boat boat_{};
+
+public:
+
+};
